Feature/oraclize funding flow

+++ Contracts/contracts/InitialVariables.sol
@@ -20,20 +20,20 @@ external  {
   database.setAddress(keccak256("MyBitFoundation"), _myBitFoundation);
   database.setAddress(keccak256("InstallerEscrow"), _installerEscrow); 
   // --------------------Asset Creation Variables-----------------
-  database.setUint(keccak256("myBitFoundationPercentage"), 1);
-  database.setUint(keccak256("stakedTokenPercentage"), 2);
-  database.setUint(keccak256("installerPercentage"), 97);
+  database.setUint(keccak256("myBitFoundationPercentage"), uint(1));
+  database.setUint(keccak256("stakedTokenPercentage"), uint(2));
+  database.setUint(keccak256("installerPercentage"), uint(97));
   // ---------------------Access Price in USD--------------------------
-  database.setUint(keccak256("accessTokenFee", uint(2)), 25);
-  database.setUint(keccak256("accessTokenFee", uint(3)), 75); 
-  database.setUint(keccak256("accessTokenFee", uint(4)), 100); 
+  database.setUint(keccak256("accessTokenFee", uint(2)), uint(25));
+  database.setUint(keccak256("accessTokenFee", uint(3)), uint(75)); 
+  database.setUint(keccak256("accessTokenFee", uint(4)), uint(100)); 
 
 
   // --------------------Operator Cost-----------------------
-  database.setUint(keccak256("operatorEscrowPercentage"), 10); 
+  database.setUint(keccak256("operatorEscrowPercentage"), uint(10)); 
 
   // -------------Oracle Variables-------------------------
-  database.setUint(keccak256("priceUpdateTimeline"), 1000);     // TODO: testing number. This is the length of time an Oracle price is valid for
+  database.setUint(keccak256("priceUpdateTimeline"), uint(1000));     // TODO: testing number. This is the length of time an Oracle price is valid for
   
   LogInitialized(msg.sender, address(database), block.number);
 }
